[lists.base]: Remove redundant save_activated()` method
This is an redundant code snippet that I used to use to emit the active
changed signal but no longer relevant.

Instead, we can call `set_active` directly in the activate method
beforethe `activeChanged` is emited. This feels like a better approach
because we want the active item to be already set correctly before we
start updating models.