Update active item after model data set swap
Each cached model data set might have an active item set. I didn't take
this into account and consequently when we're swapping the data sets the
active item was not updated leaving the models in an inconsistent state.

Making it clearer how the active flag is set seem like the right thing
to do. This is now done entirely in the model (not the view) and the
model takes care of calling the right update signals and when the new
`set_active()` method is called.