Make source code Python 3 exclusive
Bookmarks source code has been updated to drop Python 2 in favor of
Python 3. This is to bring it inline with the latest VFX platform
specifications and to make the project future proof.

Also, I'm busy and don't have the time required to support older
python versions, but perhaps this is okay as DCCs slowly migrate
over to Python 3.

In the process of updating I also started to rearrange the module
structure, moving the `rv`, `slack` and `ffmpeg` modules to a `external`
submodule and rename `bookmark_db` to `database`.