UI improvements and bug fixes
A mix of tweaks across the whole project - mostly been tweaking the UI
elements (I finally managed to clean the icons up). Whilst there I also
fixed a few delegate drawing errors and a optimized the `get_rsc_pixmap`
method so that it doesn't call the filesystem needlessly.

To be continued...