Refactor user settings file structure
The user settings file was one of the first things created for this project,
and it was never really managed well structure-wise.

This commit is an effort to reorganise how values are organised and stored
so the file can be read more easily. By the process inevitable inconsistencies
manifested, and so I tried to make sure item filter values and item selections
are stored and retrieved via consistent interfaces. That also led to other
changes. I also fixed bugs and added missing functionality in the process.

This is a comprehensive commit so bugs are probably plentiful. I'll fix them
in the near future :)