Add publish functionality.
The publishing functionality was already present, but as part of an experimental
site-specific module. I took some effort to yank it out and implement it in a more
general manner with more complete app integration.

This commit adds a new 'publish.py' module that contains the publish widget used
to publish elements. In the current implementation the publish process involves
copying items from the files tab to a publish folder using a generalized file name.

The publish file name and path can be customized via the bookmark editor and the
values are stored in the tokens module. I also made some changes how movies are
created - I'm now using OpenImageIO to create a preview jpeg copy of image sequences
and in turn this I can use to create videos. This allows the creation of video previews
of formats that ffmpeg would not support natively.