Various editor improvements and bug fixes
Whilst working on the docs I noticed some errors in the editor module.
For instance, the thumbnail editor wasn't setting the images source correctly.
On closer inspection I realised there was no need to set the source, and instead
we can use the editor's db_source to query the current thumbnail file.

I changed some out of date label text on the view module to make it more succinct.

I also made some cosmetic changes to the bookmarker widget and removed the status widget.
This included in the end tweaking the base ListWidget and ListView classes to add a
thumbnail background to them.