Various functionality and UI changes
I've made changes to the delegate - chiefly changes to how items are displayed and how
text segments and painter paths are cached and retrieved. This has a lot of benefits, a marginal
performance increase and a somewhat more organised paint event. The default appearance of
the items have been changed to make use of clickable labels - and to make bookmarks and asset items
more legible. The default font of the file items changed too.

There's a slew of other small and big code changes I didn't commit individually. This is a bad habit, I should educate myself out of!