Change QListView to QTableView
This work in progress commit changes the default view and models
to tables. This is to allow adding extra item information in the resulting new columns.

The change required tweaks to the delegate view and models. A (good)
side-effect of the tweaks were that I took a look at how the description
editor was implemented. It's one of the oldest pieces of code in the project
and as such it was incorrect. Previously the editor itself implemented all
the delegate editor functions. These have been moved into the delegate now and the delegate editor module was removed.