Add 'progress' column to the database
The progress data is a dict object we're serializing to json to store. An
issue I encountered is that the int keys of the dict are not preserved
by the json dump/loads methods. The solution is to pass an extra object_hook to convert int keys back to their original int value.