This commit adds database connection changes.
When refreshing the bookmark items model all database connectors reset.
Also, the database interface now connects using the built-in timeout and
sets a valid or invalid state depending on if the connection was successfully
made. When the connector is in invalid state no value will be read or
set to the database and the item delegate will visually indicate the
invalid state.