Implement bookmark and asset item display tokens
This is a more comprehensive commit to address the display of asset and bookmark items.
The main issue is as Bookmarks wraps around a pipeline the display names can resolve into
long and confusing names. For instance, if a bookmark item's path is
`\\server\projects\WIP\Client\Job\080_Episodes\EpisodeA\Sequences`
what should be the display name of the bookmark item inside bookmarks? There are various
correct answers but we can probably agree that "server/projects/WIP", and "080_Episodes" might be
superflous strings. So instead, we can use the tokens module and assign a display token to the item,
such as`{job1}/{root1}/{root2}` resulting in `Job/EpisodeA/Sequences`, a more user-friendly name.

The resulting changes also called for changes in the delegate module and more fine grained user control
over what and how is being displayed.