Script menu changes
Tweaks to add support for separators for the Scripts menus. Also, made the Maya scripts menu
behave the same way as the main Scripts menus but the code is duplicated at the moment.

I'd like to make another pass on this at a later stage to add in the options to display scripts
per job or root values. Also, is there any merit to controlling the source of the scripts? E.g.
via an environment value, e.g. "BOOKMARKS_SCRIPTS_ROOT", and "BOOKMARKS_MAYA_SCRIPTS_ROOT"?