Add ShotGrid entity and task filter menus
This commit adds two new filter setter ui menus to the top bar to set the
ShotGrid entity OR the ShotGrid task.

The process involved a cleanup pass on some of the ui and backend. I fixed some bugs
related to how the model's internal data was sorted previously and this made me realise
the DataDict implementation had a bug - it wasn't copying over all properties correctly.

The new filter menus values are populated by the worker threads and are saved into the main
DataDict instances the filter combo box models read.