Fix RV push url handler not ready when pushing footage to RV
The RV url handler does not work if the RV session doesn't have the handler
ready. This happens every time when a fresh instance is launched. I'm trying to
balance this by launching a dummy python command, wait for the process to run, then
a litte more, THEN push footages to RV.

More testing is needed, but seems to work at first glance.