Implement Asset Links and Template api and ui (WIP).
Having not worked on Bookmarks in some time, I wanted to revisit old code - and change how size and color constants are defined. There's a lost of tweaks outside of the new modules - stylesheet has been tweaked and there's a plethora of little incremental changes.

This work is still in progress...